+++ grails-app/services/com/softamo/movilrural/UploadVillageFeaturedImageService.groovy
@@ -9,13 +9,13 @@ class UploadVillageFeaturedImageService implements ImageUploadable {
     VillageGormService villageGormService
 
     Village uploadFeatureImage(FeaturedImageCommand cmd) {
-        uploadImageToS3(cmd.id, cmd.version, cmd.featuredImageFile) { Long id, Integer version, String s3FileUrl ->
+        uploadImageToS3(cmd.id, cmd.version, cmd.featuredImageFile) { Long id, Long version, String s3FileUrl ->
             villageGormService.updateFeaturedImageUrl(id, version, s3FileUrl)
         } as Village
     }
 
     Village uploadImage(ImageCommand cmd) {
-        uploadImageToS3(cmd.id, cmd.version, cmd.imageFile) { Long id, Integer version, String s3FileUrl ->
+        uploadImageToS3(cmd.id, cmd.version, cmd.imageFile) { Long id, Long version, String s3FileUrl ->
             villageGormService.addImageUrl(cmd.id, cmd.version, s3FileUrl)
         } as Village
     }
